Fix: tela do ADM (ref#57)

+++ seth/src/components/admin/estacoes/Funcionalidades/DeletarEstacao.tsx
@@ -1,22 +1,66 @@
-import { useState } from "react"
+import { useEffect, useState } from "react"
 import axios from "axios"
+import { Estacao } from "../../../../types/estacao"
+import { Parametro } from "../../../../types/parametro"
+import "../../main.css";
 
 
 export default function DeleteEstacao() {
 
-    const[id, setId] = useState('')
+    const[id, setId] = useState(0)
+    const [estacoes, setEstacoes] = useState<Array<Estacao>>([])
+    const [parametros, setParametros] = useState<Array<Parametro>>([])
 
     function deletar() {
-        axios.delete(`http://localhost:30105/api/estacao/deletar/${id}`)
-        .then(() =>{
-            alert("Estação deletada com sucesso!")
-            
-            
+
+        if (estacoes.find((estacao) => estacao.cod_estacao === id) != undefined) {
+
+            const estacao = estacoes.find((estacao) => estacao.cod_estacao === id)
+
+            const parametro = parametros.find((parametro) => parametro.cod_estacao === id)
+
+
+
+            if (parametro?.cod_estacao === estacao?.cod_estacao) {
+
+                alert("Não é possível deletar uma estação que possui um parâmetro cadastrado!")
+                
+            } else {
+
+                axios.delete(`http://localhost:30105/api/estacao/deletar/${id}`)
+                .then(() =>{
+                    alert("Estação deletada com sucesso!")
+                    
+                    
+                })
+                .catch((error) =>{
+                    console.error(error)
+                })
+            }
+        }
+        
+
+
+    }
+
+
+    useEffect(() => {
+        axios.get('http://localhost:30105/api/estacoes')
+        .then((response) => {
+            setEstacoes(response.data)
         })
-        .catch((error) =>{
+        .catch((error) => {
             console.error(error)
         })
-    }
+
+        axios.get(`http://localhost:30105/api/parametros`)
+        .then((response) => {
+            setParametros(response.data)
+        })
+        .catch((error) => {
+            console.error(error)
+        })
+    }, [])
 
     return(
         <>
@@ -33,8 +77,19 @@ export default function DeleteEstacao() {
                 <div id="Inputs_Camp">
 
                     <p>
-                        ID (Somente Número(s) ): 
-                        <input type="number" value={id} onChange={(event)=>setId(event.target.value)} /> 
+                        <select name="" value={id} onChange={(event) => setId(parseInt(event.target.value))} id="Select_Estacao">
+
+                            <option value="">Selecione uma Estação</option>
+
+                            {estacoes.map((estacao) => {
+                                return(
+                                    <option key={estacao.cod_estacao} value={estacao.cod_estacao}>
+                                        {estacao.nome}
+                                    </option>
+                                )
+                            })}
+
+                        </select>
                     </p>
 
                 </div>
